split up ir module
