+++ fluree-db-query/src/ir/projection.rs
@@ -1,0 +1,338 @@
+//! Query output shape: the top-level `Query` plus the selection specs that
+//! describe how results are projected and crawled into nested JSON-LD form.
+
+use super::pattern::Pattern;
+use crate::triple::TriplePattern;
+use crate::var_registry::VarId;
+use fluree_db_core::Sid;
+
+/// Top-level query structure
+///
+/// Represents a parsed and lowered query ready for planning.
+#[derive(Debug, Clone)]
+pub struct Query {
+    /// Selected variables (output columns)
+    pub select: Vec<VarId>,
+    /// Where clause patterns (ordered!)
+    pub where_: Vec<Pattern>,
+}
+
+impl Query {
+    /// Create a new query
+    pub fn new(select: Vec<VarId>, where_: Vec<Pattern>) -> Self {
+        Self { select, where_ }
+    }
+
+    /// Create a query with a single triple pattern
+    pub fn single(select: Vec<VarId>, pattern: TriplePattern) -> Self {
+        Self {
+            select,
+            where_: vec![Pattern::Triple(pattern)],
+        }
+    }
+
+    /// Get all triple patterns in the where clause (flattening nested structures)
+    pub fn triple_patterns(&self) -> Vec<&TriplePattern> {
+        fn collect<'a>(patterns: &'a [Pattern], out: &mut Vec<&'a TriplePattern>) {
+            for p in patterns {
+                match p {
+                    Pattern::Triple(tp) => out.push(tp),
+                    Pattern::Optional(inner)
+                    | Pattern::Minus(inner)
+                    | Pattern::Exists(inner)
+                    | Pattern::NotExists(inner) => collect(inner, out),
+                    Pattern::Union(branches) => {
+                        for branch in branches {
+                            collect(branch, out);
+                        }
+                    }
+                    Pattern::Graph { patterns, .. } => collect(patterns, out),
+                    Pattern::Service(sp) => collect(&sp.patterns, out),
+                    Pattern::Filter(_)
+                    | Pattern::Bind { .. }
+                    | Pattern::Values { .. }
+                    | Pattern::PropertyPath(_)
+                    | Pattern::Subquery(_)
+                    | Pattern::IndexSearch(_)
+                    | Pattern::VectorSearch(_)
+                    | Pattern::R2rml(_)
+                    | Pattern::GeoSearch(_)
+                    | Pattern::S2Search(_) => {}
+                }
+            }
+        }
+
+        let mut result = Vec::new();
+        collect(&self.where_, &mut result);
+        result
+    }
+}
+
+// ============================================================================
+// Graph crawl select types (resolved)
+// ============================================================================
+
+/// Root of a graph crawl select (resolved)
+///
+/// Supports both variable and IRI constant roots:
+/// - Variable root: from query results (e.g., `?person`)
+/// - IRI constant root: direct subject fetch (e.g., `ex:alice`)
+#[derive(Debug, Clone, PartialEq)]
+pub enum Root {
+    /// Variable root - value comes from query results
+    Var(VarId),
+    /// IRI constant root - direct subject lookup
+    Sid(Sid),
+}
+
+/// Nested selection specification for sub-crawls (resolved)
+///
+/// This type captures the full selection state for nested property expansion,
+/// including both forward and reverse properties.
+#[derive(Debug, Clone, PartialEq)]
+pub struct NestedSelectSpec {
+    /// Forward property selections
+    pub forward: Vec<SelectionSpec>,
+    /// Reverse property selections (predicate Sid → optional nested spec)
+    /// None means no sub-selections (just return @id), Some means nested expansion
+    pub reverse: std::collections::HashMap<Sid, Option<Box<NestedSelectSpec>>>,
+    /// Whether wildcard was specified at this level
+    pub has_wildcard: bool,
+}
+
+impl NestedSelectSpec {
+    /// Create a new nested select spec
+    pub fn new(
+        forward: Vec<SelectionSpec>,
+        reverse: std::collections::HashMap<Sid, Option<Box<NestedSelectSpec>>>,
+        has_wildcard: bool,
+    ) -> Self {
+        Self {
+            forward,
+            reverse,
+            has_wildcard,
+        }
+    }
+
+    /// Check if this spec is empty (no selections)
+    pub fn is_empty(&self) -> bool {
+        self.forward.is_empty() && self.reverse.is_empty()
+    }
+}
+
+/// Selection specification for graph crawl (resolved)
+///
+/// Defines what properties to include at each level of expansion.
+#[derive(Debug, Clone, PartialEq)]
+pub enum SelectionSpec {
+    /// Explicit @id selection (include @id even when wildcard is not specified)
+    Id,
+    /// Wildcard - select all properties at this level
+    Wildcard,
+    /// Property selection with optional nested expansion
+    Property {
+        /// Predicate Sid
+        predicate: Sid,
+        /// Optional nested selection spec for expanding this property's values
+        /// Uses Box to avoid infinite type recursion
+        sub_spec: Option<Box<NestedSelectSpec>>,
+    },
+}
+
+/// Graph crawl selection specification (resolved, with Sids)
+///
+/// This is the resolved form of `UnresolvedGraphSelectSpec`, with all IRIs
+/// encoded as Sids. Used during result formatting for nested JSON-LD output.
+#[derive(Debug, Clone, PartialEq)]
+pub struct GraphSelectSpec {
+    /// Root of the crawl - variable or IRI constant
+    pub root: Root,
+    /// Forward property selections
+    pub selections: Vec<SelectionSpec>,
+    /// Reverse property selections (predicate Sid → optional nested spec)
+    /// None means no sub-selections (just return @id), Some means nested expansion
+    pub reverse: std::collections::HashMap<Sid, Option<Box<NestedSelectSpec>>>,
+    /// Max depth for auto-expansion (0 = no auto-expand)
+    pub depth: usize,
+    /// Whether wildcard was specified (controls @id inclusion)
+    pub has_wildcard: bool,
+}
+
+impl GraphSelectSpec {
+    /// Create a new graph select spec
+    pub fn new(root: Root, selections: Vec<SelectionSpec>) -> Self {
+        let has_wildcard = selections
+            .iter()
+            .any(|s| matches!(s, SelectionSpec::Wildcard));
+        Self {
+            root,
+            selections,
+            reverse: std::collections::HashMap::new(),
+            depth: 0,
+            has_wildcard,
+        }
+    }
+
+    /// Returns the root variable, if the root is a variable (not an IRI constant).
+    pub fn root_var(&self) -> Option<VarId> {
+        match &self.root {
+            Root::Var(v) => Some(*v),
+            Root::Sid(_) => None,
+        }
+    }
+
+    /// Generate a hash for cache keying purposes
+    ///
+    /// Used to differentiate the same Sid expanded under different specs.
+    /// The cache key is `(Sid, spec_hash, depth_remaining)`.
+    pub fn spec_hash(&self) -> u64 {
+        use std::hash::{Hash, Hasher};
+        let mut hasher = std::collections::hash_map::DefaultHasher::new();
+
+        self.has_wildcard.hash(&mut hasher);
+        self.selections.len().hash(&mut hasher);
+
+        // Hash selection structure
+        fn hash_selection(spec: &SelectionSpec, hasher: &mut impl Hasher) {
+            match spec {
+                SelectionSpec::Id => {
+                    2u8.hash(hasher);
+                }
+                SelectionSpec::Wildcard => {
+                    0u8.hash(hasher);
+                }
+                SelectionSpec::Property {
+                    predicate,
+                    sub_spec,
+                } => {
+                    1u8.hash(hasher);
+                    predicate.hash(hasher);
+                    if let Some(nested) = sub_spec {
+                        // Hash nested spec
+                        nested.has_wildcard.hash(hasher);
+                        nested.forward.len().hash(hasher);
+                        for sub in &nested.forward {
+                            hash_selection(sub, hasher);
+                        }
+                        // Hash nested reverse
+                        nested.reverse.len().hash(hasher);
+                        let mut nested_rev_keys: Vec<_> = nested.reverse.keys().collect();
+                        nested_rev_keys.sort();
+                        for key in nested_rev_keys {
+                            key.hash(hasher);
+                            if let Some(nested_nested) = nested.reverse.get(key) {
+                                if let Some(spec) = nested_nested {
+                                    1u8.hash(hasher);
+                                    hash_nested_spec(spec, hasher);
+                                } else {
+                                    0u8.hash(hasher);
+                                }
+                            }
+                        }
+                    } else {
+                        0usize.hash(hasher);
+                    }
+                }
+            }
+        }
+
+        fn hash_nested_spec(spec: &NestedSelectSpec, hasher: &mut impl Hasher) {
+            spec.has_wildcard.hash(hasher);
+            spec.forward.len().hash(hasher);
+            for sub in &spec.forward {
+                hash_selection(sub, hasher);
+            }
+            spec.reverse.len().hash(hasher);
+            let mut rev_keys: Vec<_> = spec.reverse.keys().collect();
+            rev_keys.sort();
+            for key in rev_keys {
+                key.hash(hasher);
+                if let Some(nested) = spec.reverse.get(key) {
+                    if let Some(inner) = nested {
+                        1u8.hash(hasher);
+                        hash_nested_spec(inner, hasher);
+                    } else {
+                        0u8.hash(hasher);
+                    }
+                }
+            }
+        }
+
+        for sel in &self.selections {
+            hash_selection(sel, &mut hasher);
+        }
+
+        // Hash reverse properties
+        self.reverse.len().hash(&mut hasher);
+        // Sort by Sid for deterministic hashing
+        let mut reverse_keys: Vec<_> = self.reverse.keys().collect();
+        reverse_keys.sort();
+        for key in reverse_keys {
+            key.hash(&mut hasher);
+            if let Some(nested) = self.reverse.get(key) {
+                if let Some(spec) = nested {
+                    1u8.hash(&mut hasher);
+                    hash_nested_spec(spec, &mut hasher);
+                } else {
+                    0u8.hash(&mut hasher);
+                }
+            }
+        }
+
+        hasher.finish()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::triple::{Ref, Term};
+
+    fn test_pattern() -> TriplePattern {
+        TriplePattern::new(
+            Ref::Var(VarId(0)),
+            Ref::Sid(Sid::new(100, "name")),
+            Term::Var(VarId(1)),
+        )
+    }
+
+    #[test]
+    fn test_query_new() {
+        let pattern = test_pattern();
+        let query = Query::new(vec![VarId(0), VarId(1)], vec![Pattern::Triple(pattern)]);
+
+        assert_eq!(query.select.len(), 2);
+        assert_eq!(query.where_.len(), 1);
+    }
+
+    #[test]
+    fn test_query_single() {
+        let pattern = test_pattern();
+        let query = Query::single(vec![VarId(0)], pattern);
+
+        assert_eq!(query.where_.len(), 1);
+        assert!(query.where_[0].is_triple());
+    }
+
+    #[test]
+    fn test_query_triple_patterns() {
+        let p1 = TriplePattern::new(
+            Ref::Var(VarId(0)),
+            Ref::Sid(Sid::new(100, "name")),
+            Term::Var(VarId(1)),
+        );
+        let p2 = TriplePattern::new(
+            Ref::Var(VarId(0)),
+            Ref::Sid(Sid::new(101, "age")),
+            Term::Var(VarId(2)),
+        );
+
+        let query = Query::new(
+            vec![VarId(0), VarId(1), VarId(2)],
+            vec![Pattern::Triple(p1), Pattern::Triple(p2)],
+        );
+
+        let triples = query.triple_patterns();
+        assert_eq!(triples.len(), 2);
+    }
+}
